+++ example/lib/injector/Service.dart
@@ -1,0 +1,5 @@
+import 'package:injectable/injectable.dart';
+
+abstract class Service {
+  Set<String> get environments;
+}