+++ example/test/test_service.dart
@@ -2,7 +2,5 @@ import 'package:injectable/injectable.dart';
 
 @injectable
 class TestService {
-  final String xd;
-
-  TestService(this.xd);
+  TestService();
 }