+++ injectable_generator/lib/generator/module_factory_generator.dart
@@ -5,17 +5,15 @@ class ModuleFactoryGenerator extends RegisterFuncGenerator {
   ModuleFactoryGenerator(Set<ImportableType> prefixedTypes) : super(prefixedTypes);
 
   @override
-  String generateInitializer(DependencyConfig dep, {String getIt = 'g'}) {
+  String generateInitializer(DependencyConfig dep, {String getIt = 'get'}) {
     final flattenedParams = flattenParams(dep.dependencies, getIt);
-
     final constructorName = dep.constructorName != null && dep.constructorName.isNotEmpty ? '.${dep.constructorName}' : '';
-
-    return '${dep.typeImpl.getDisplayName(prefixedTypes, withTypeArgs: false)}$constructorName($flattenedParams)';
+    return '${dep.typeImpl.getDisplayName(prefixedTypes, includeTypeArgs: false)}$constructorName($flattenedParams)';
   }
 
   @override
   String generate(DependencyConfig dep, {Set<ImportableType> prefixedTypes}) {
-    final constructor = generateInitializer(dep, getIt: '_g');
+    final constructor = generateInitializer(dep, getIt: '_get');
     if (dep.isModuleMethod) {
       return '${dep.typeImpl.getDisplayName(prefixedTypes)} ${dep.initializerName}() => $constructor;';
     } else {